+++ dump_things_service/tests/test_extract_inline.py
@@ -58,21 +58,21 @@ class MockedModule:
 
 def get_inlined_object(model_module):
     return model_module.Person(
-        pid='trr379:test_extract_1',
+        pid='dlflatsocial:test_extract_1',
         given_name='Grandfather',
         relations={
-            'trr379:test_extract_1_1': model_module.Person(
-                pid='trr379:test_extract_1_1',
+            'dlflatsocial:test_extract_1_1': model_module.Person(
+                pid='dlflatsocial:test_extract_1_1',
                 given_name='Father',
                 relations={
-                    'trr379:test_extract_1_1_1': model_module.Agent(
-                        pid='trr379:test_extract_1_1_1',
-                        acted_on_behalf_of=['trr379:test_extract_1_1'],
+                    'dlflatsocial:test_extract_1_1_1': model_module.Agent(
+                        pid='dlflatsocial:test_extract_1_1_1',
+                        acted_on_behalf_of=['dlflatsocial:test_extract_1_1'],
                     ),
                 },
             ),
-            'trr379:test_extract_1_2': model_module.InstantaneousEvent(
-                pid='trr379:test_extract_1_2',
+            'dlflatsocial:test_extract_1_2': model_module.InstantaneousEvent(
+                pid='dlflatsocial:test_extract_1_2',
                 at_time='2028-12-31',
             ),
         },
@@ -84,12 +84,12 @@ inlined_json_record = cleaned_json(dataclasses.asdict(inlined_object))
 
 
 empty_inlined_object = Person(
-    pid='trr379:test_extract_a',
+    pid='dlflatsocial:test_extract_a',
     given_name='Opa',
     relations={
-        'trr379:test_extract_a_a': Thing(pid='trr379:test_extract_a_a'),
-        'trr379:test_extract_a_b': Thing(pid='trr379:test_extract_a_b'),
-        'trr379:test_extract_a_c': Thing(pid='trr379:test_extract_a_c'),
+        'dlflatsocial:test_extract_a_a': Thing(pid='dlflatsocial:test_extract_a_a'),
+        'dlflatsocial:test_extract_a_b': Thing(pid='dlflatsocial:test_extract_a_b'),
+        'dlflatsocial:test_extract_a_c': Thing(pid='dlflatsocial:test_extract_a_c'),
     },
 )
 
@@ -97,49 +97,49 @@ empty_inlined_json_record = cleaned_json(dataclasses.asdict(empty_inlined_object
 
 
 tree = (
-    ('trr379:test_extract_1', ('trr379:test_extract_1_1', 'trr379:test_extract_1_2')),
-    ('trr379:test_extract_1_1', ('trr379:test_extract_1_1_1',)),
-    ('trr379:test_extract_1_2', ()),
-    ('trr379:test_extract_1_1_1', ()),
+    ('dlflatsocial:test_extract_1', ('dlflatsocial:test_extract_1_1', 'dlflatsocial:test_extract_1_2')),
+    ('dlflatsocial:test_extract_1_1', ('dlflatsocial:test_extract_1_1_1',)),
+    ('dlflatsocial:test_extract_1_2', ()),
+    ('dlflatsocial:test_extract_1_1_1', ()),
 )
 
 
 ttl_with_inline_a = """
-@prefix dlsocial: <https://concepts.datalad.org/s/social/unreleased/> .
+@prefix dlsocialmx: <https://concepts.datalad.org/s/social-mixin/unreleased/> .
 @prefix dlthings: <https://concepts.datalad.org/s/things/v1/> .
-@prefix trr379: <https://trr379.de/> .
+@prefix dlflatsocial <https://concepts.datalad.org/s/flat-social/unreleased/> .
 
-trr379:test_ttl_inline_1 a dlsocial:Person ;
-    dlsocial:given_name "Grandfather" ;
-    dlthings:relation trr379:test_ttl_inline_1_1,
-        trr379:test_ttl_inline_1_2 .
+dlflatsocial:test_ttl_inline_1 a dlsocial:Person ;
+    dlsocialmx:given_name "Grandfather" ;
+    dlthings:relation dlflatsocial:test_ttl_inline_1_1,
+        dlflatsocial:test_ttl_inline_1_2 .
 """
 
 ttl_with_inline_b = """
 @prefix dlprov: <https://concepts.datalad.org/s/prov/unreleased/> .
-@prefix trr379: <https://trr379.de/> .
+@prefix dlflatsocial <https://trr379.de/> .
 
-trr379:test_ttl_inline_1_1_1 a dlprov:Agent ;
-    dlprov:acted_on_behalf_of trr379:test_ttl_inline_1_1 .
+dlflatsocial:test_ttl_inline_1_1_1 a dlprov:Agent ;
+    dlprov:acted_on_behalf_of dlflatsocial:test_ttl_inline_1_1 .
 """
 
 ttl_with_inline_c = """
 @prefix dltemporal: <https://concepts.datalad.org/s/temporal/unreleased/> .
-@prefix trr379: <https://trr379.de/> .
+@prefix dlflatsocial <https://concepts.datalad.org/s/flat-social/unreleased/> .
 @prefix w3ctr: <https://www.w3.org/TR/> .
 
-trr379:test_ttl_inline_1_2 a dltemporal:InstantaneousEvent ;
+dlflatsocial:test_ttl_inline_1_2 a dltemporal:InstantaneousEvent ;
     dltemporal:at_time "2028-12-31"^^w3ctr:NOTE-datetime .
 """
 
 ttl_with_inline_d = """
-@prefix dlsocial: <https://concepts.datalad.org/s/social/unreleased/> .
+@prefix dlsocialmx: <https://concepts.datalad.org/s/social-mixin/unreleased/> .
 @prefix dlthings: <https://concepts.datalad.org/s/things/v1/> .
-@prefix trr379: <https://trr379.de/> .
+@prefix dlflatsocial <https://trr379.de/> .
 
-trr379:test_ttl_inline_1_1 a dlsocial:Person ;
-    dlsocial:given_name "Father" ;
-    dlthings:relation trr379:test_ttl_inline_1_1_1 .
+dlflatsocial:test_ttl_inline_1_1 a dlsocial:Person ;
+    dlsocialmx:given_name "Father" ;
+    dlthings:relation dlflatsocial:test_ttl_inline_1_1_1 .
 
 """
 
@@ -152,12 +152,12 @@ ttls_with_inline = (
 
 ttl_tree = (
     (
-        'trr379:test_ttl_inline_1',
-        ('trr379:test_ttl_inline_1_1', 'trr379:test_ttl_inline_1_2'),
+        'dlflatsocial:test_ttl_inline_1',
+        ('dlflatsocial:test_ttl_inline_1_1', 'dlflatsocial:test_ttl_inline_1_2'),
     ),
-    ('trr379:test_ttl_inline_1_1', ('trr379:test_ttl_inline_1_1_1',)),
-    ('trr379:test_ttl_inline_1_2', ()),
-    ('trr379:test_ttl_inline_1_1_1', ()),
+    ('dlflatsocial:test_ttl_inline_1_1', ('dlflatsocial:test_ttl_inline_1_1_1',)),
+    ('dlflatsocial:test_ttl_inline_1_2', ()),
+    ('dlflatsocial:test_ttl_inline_1_1_1', ()),
 )
 
 
@@ -200,17 +200,20 @@ def test_dont_extract_empty_things_locally():
     assert records[0] == empty_inlined_object
 
 
+# We skip this test because the dlflatsocial-schema does not support inlined
+# relations
+@pytest.mark.xfail
 def test_inline_extraction_on_service(fastapi_client_simple):
     test_client, _ = fastapi_client_simple
 
     for i in range(1, 3):
         # Deposit JSON record
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person',
+            f'/collection_dlflatsocial-{i}/record/Person',
             headers={'x-dumpthings-token': 'token_1'},
             json=inlined_json_record,
         )
-        assert response.status_code == HTTP_200_OK
+        assert response.status_code == HTTP_200_OK, 'Response content:' + response.text
         # Check linkage between records
         _check_result_json(response.json(), tree)
 
@@ -219,7 +222,7 @@ def test_inline_extraction_on_service(fastapi_client_simple):
         records = []
         for record_pid in (entry[0] for entry in tree):
             response = test_client.get(
-                f'/collection_trr379-{i}/record?pid={record_pid}',
+                f'/collection_dlflatsocial-{i}/record?pid={record_pid}',
                 headers={'x-dumpthings-token': 'token_1'},
             )
             assert response.status_code == HTTP_200_OK
@@ -230,18 +233,20 @@ def test_inline_extraction_on_service(fastapi_client_simple):
 
         # Check that individual record classes were recognized
         for class_name, pids in (
-            ('Person', ('trr379:test_extract_1', 'trr379:test_extract_1_1')),
-            ('Agent', ('trr379:test_extract_1_1_1',)),
-            ('InstantaneousEvent', ('trr379:test_extract_1_2',)),
+            ('Person', ('dlflatsocial:test_extract_1', 'dlflatsocial:test_extract_1_1')),
+            ('Agent', ('dlflatsocial:test_extract_1_1_1',)),
+            ('InstantaneousEvent', ('dlflatsocial:test_extract_1_2',)),
         ):
             records = test_client.get(
-                f'/collection_trr379-{i}/records/{class_name}',
+                f'/collection_dlflatsocial-{i}/records/{class_name}',
                 headers={'x-dumpthings-token': 'token_1'},
             ).json()
             for pid in pids:
                 assert any(record['pid'] == pid for record in records)
 
 
+# We skip this test because the dlflatsocial-schema does not support inlined
+# relations
 @pytest.mark.xfail
 def test_inline_ttl_processing(fastapi_client_simple):
     test_client, _ = fastapi_client_simple
@@ -250,7 +255,7 @@ def test_inline_ttl_processing(fastapi_client_simple):
         # Deposit TTL records
         for class_name, ttl_record in ttls_with_inline:
             response = test_client.post(
-                f'/collection_trr379-{i}/record/{class_name}?format=ttl',
+                f'/collection_dlflatsocial-{i}/record/{class_name}?format=ttl',
                 headers={'x-dumpthings-token': 'token_1'},
                 json=ttl_record,
             )
@@ -261,7 +266,7 @@ def test_inline_ttl_processing(fastapi_client_simple):
         records = []
         for record_pid in (entry[0] for entry in ttl_tree):
             response = test_client.get(
-                f'/collection_trr379-{i}/record?pid={record_pid}',
+                f'/collection_dlflatsocial-{i}/record?pid={record_pid}',
                 headers={'x-dumpthings-token': 'token_1'},
             )
             assert response.status_code == HTTP_200_OK
@@ -272,12 +277,12 @@ def test_inline_ttl_processing(fastapi_client_simple):
 
         # Check that individual record classes were recognized
         for class_name, pids in (
-            ('Person', ('trr379:test_ttl_inline_1', 'trr379:test_ttl_inline_1_1')),
-            ('Agent', ('trr379:test_ttl_inline_1_1_1',)),
-            ('InstantaneousEvent', ('trr379:test_ttl_inline_1_2',)),
+            ('Person', ('dlflatsocial:test_ttl_inline_1', 'dlflatsocial:test_ttl_inline_1_1')),
+            ('Agent', ('dlflatsocial:test_ttl_inline_1_1_1',)),
+            ('InstantaneousEvent', ('dlflatsocial:test_ttl_inline_1_2',)),
         ):
             records = test_client.get(
-                f'/collection_trr379-{i}/records/{class_name}',
+                f'/collection_dlflatsocial-{i}/records/{class_name}',
                 headers={'x-dumpthings-token': 'token_1'},
             ).json()
             for pid in pids:
@@ -311,7 +316,7 @@ def test_dont_extract_empty_things_on_service(fastapi_client_simple):
     for i in range(1, 3):
         # Deposit JSON record
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person',
+            f'/collection_dlflatsocial-{i}/record/Person',
             headers={'x-dumpthings-token': 'token_1'},
             json=empty_inlined_json_record,
         )