+++ dump_things_service/tests/test_roundtrip_flatsocial.py
@@ -4,57 +4,57 @@ from .. import HTTP_200_OK
 from ..utils import cleaned_json
 
 json_record = {
-    'pid': 'trr379:test_john_json',
+    'pid': 'dlflatsocial:test_john_json',
     'given_name': 'Johnöüß',
 }
 json_record_out = {
-    'schema_type': 'dlsocial:Person',
+    'schema_type': 'dlflatsocial:Person',
     **json_record,
 }
 
-new_ttl_pid = 'trr379:another_john_json'
+new_ttl_pid = 'dlflatsocial:another_john_json'
 
-ttl_record = """@prefix dlsocial: <https://concepts.datalad.org/s/social/unreleased/> .
+ttl_record = """@prefix dlflatsocial: <https://concepts.datalad.org/s/flat-social/unreleased/> .
+@prefix dlsocialmx: <https://concepts.datalad.org/s/social-mixin/unreleased/> .
 @prefix dlthings: <https://concepts.datalad.org/s/things/v1/> .
-@prefix trr379: <https://trr379.de/> .
 
-trr379:test_john_ttl a dlsocial:Person ;
-    dlsocial:given_name "Johnöüß" ;
+dlflatsocial:test_john_ttl a dlflatsocial:Person ;
+    dlsocialmx:given_name "Johnöüß" ;
     dlthings:annotations [ a dlthings:Annotation ;
             dlthings:annotation_tag <http://purl.obolibrary.org/obo/NCIT_C54269> ;
             dlthings:annotation_value "test_user_1" ] .
 """
 
-ttl_input_record = """@prefix dlsocial: <https://concepts.datalad.org/s/social/unreleased/> .
+ttl_input_record = """@prefix dlflatsocial: <https://concepts.datalad.org/s/flat-social/unreleased/> .
+@prefix dlsocialmx: <https://concepts.datalad.org/s/social-mixin/unreleased/> .
 @prefix dlthings: <https://concepts.datalad.org/s/things/v1/> .
-@prefix trr379: <https://trr379.de/> .
 
-trr379:test_john_ttl a dlsocial:Person ;
-    dlsocial:given_name "Johnöüß" .
+dlflatsocial:test_john_ttl a dlflatsocial:Person ;
+    dlsocialmx:given_name "Johnöüß" .
 """
 
-ttl_output_record = """@prefix dlsocial: <https://concepts.datalad.org/s/social/unreleased/> .
+ttl_output_record = """@prefix dlflatsocial: <https://concepts.datalad.org/s/flat-social/unreleased/> .
+@prefix dlsocialmx: <https://concepts.datalad.org/s/social-mixin/unreleased/> .
 @prefix dlthings: <https://concepts.datalad.org/s/things/v1/> .
 @prefix obo: <http://purl.obolibrary.org/obo/> .
-@prefix trr379: <https://trr379.de/> .
 
-trr379:test_john_ttl a dlsocial:Person ;
-    dlsocial:given_name "Johnöüß" ;
+dlflatsocial:test_john_ttl a dlflatsocial:Person ;
+    dlsocialmx:given_name "Johnöüß" ;
     dlthings:annotations [ a dlthings:Annotation ;
             dlthings:annotation_tag obo:NCIT_C54269 ;
             dlthings:annotation_value "test_user_1" ] .
 """
 
-new_json_pid = 'trr379:another_john_ttl'
+new_json_pid = 'dlflatsocial:another_john_ttl'
 
 
-def test_json_ttl_json_trr379(fastapi_client_simple):
+def test_json_ttl_json_dlflatsocial(fastapi_client_simple):
     test_client, _ = fastapi_client_simple
 
     for i in range(1, 3):
         # Deposit JSON records
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person',
+            f'/collection_dlflatsocial-{i}/record/Person',
             headers={'x-dumpthings-token': 'token_1'},
             json=json_record,
         )
@@ -62,7 +62,7 @@ def test_json_ttl_json_trr379(fastapi_client_simple):
 
         # Retrieve TTL records
         response = test_client.get(
-            f'/collection_trr379-{i}/record?pid={json_record["pid"]}&format=ttl',
+            f'/collection_dlflatsocial-{i}/record?pid={json_record["pid"]}&format=ttl',
             headers={'x-dumpthings-token': 'token_1'},
         )
         assert response.status_code == HTTP_200_OK
@@ -72,7 +72,7 @@ def test_json_ttl_json_trr379(fastapi_client_simple):
         ttl = ttl.replace(json_record['pid'], new_ttl_pid)
 
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person?format=ttl',
+            f'/collection_dlflatsocial-{i}/record/Person?format=ttl',
             headers={'content-type': 'text/turtle', 'x-dumpthings-token': 'token_1'},
             data=ttl,
         )
@@ -80,7 +80,7 @@ def test_json_ttl_json_trr379(fastapi_client_simple):
 
         # Retrieve JSON record
         response = test_client.get(
-            f'/collection_trr379-{i}/record?pid={new_ttl_pid}&format=json',
+            f'/collection_dlflatsocial-{i}/record?pid={new_ttl_pid}&format=json',
             headers={'x-dumpthings-token': 'token_1'},
         )
         assert response.status_code == HTTP_200_OK
@@ -90,25 +90,24 @@ def test_json_ttl_json_trr379(fastapi_client_simple):
         assert json_object == json_record_out
 
 
-@pytest.mark.xfail
-def test_ttl_json_ttl_trr379(fastapi_client_simple):
+def test_ttl_json_ttl_dlflatsocial(fastapi_client_simple):
     test_client, _ = fastapi_client_simple
 
     for i in range(1, 3):
         # Deposit a ttl record
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person?format=ttl',
+            f'/collection_dlflatsocial-{i}/record/Person?format=ttl',
             headers={
                 'x-dumpthings-token': 'token_1',
                 'content-type': 'text/turtle',
             },
             data=ttl_input_record,
         )
-        assert response.status_code == HTTP_200_OK
+        assert response.status_code == HTTP_200_OK, 'Response content: ' + response.content.decode()
 
         # Retrieve JSON records
         response = test_client.get(
-            f'/collection_trr379-{i}/record?pid=trr379:test_john_ttl&format=json',
+            f'/collection_dlflatsocial-{i}/record?pid=dlflatsocial:test_john_ttl&format=json',
             headers={'x-dumpthings-token': 'token_1'},
         )
         assert response.status_code == HTTP_200_OK
@@ -118,7 +117,7 @@ def test_ttl_json_ttl_trr379(fastapi_client_simple):
         json_object['pid'] = new_json_pid
 
         response = test_client.post(
-            f'/collection_trr379-{i}/record/Person?format=json',
+            f'/collection_dlflatsocial-{i}/record/Person?format=json',
             headers={'x-dumpthings-token': 'token_1'},
             json=json_object,
         )
@@ -126,11 +125,11 @@ def test_ttl_json_ttl_trr379(fastapi_client_simple):
 
         # Retrieve ttl record
         response = test_client.get(
-            f'/collection_trr379-{i}/record?pid={new_json_pid}&format=ttl',
+            f'/collection_dlflatsocial-{i}/record?pid={new_json_pid}&format=ttl',
             headers={'x-dumpthings-token': 'token_1'},
         )
         assert response.status_code == HTTP_200_OK
         assert (
             response.text.strip()
-            == ttl_output_record.replace('trr379:test_john_ttl', new_json_pid).strip()
+            == ttl_output_record.replace('dlflatsocial:test_john_ttl', new_json_pid).strip()
         )